importer/requirements.txt
pandas==2.2.2
mysql-connector-python==8.0.33
python-dateutil==2.8.2
